# Construct causal mask on-the-fly