# Description

Previously, we were allocating causal masks of size `(n_ctx, n_ctx)` for every instantiation of `AbstractAttention`, where `n_ctx` corresponds to the _maximum_ context length.

For models with a large maximum context length, this leads to wasteful memory consumption.

This PR 

I tried to make the change as conservatively as possible - I took the existing logic for creating the causal mask from `AbstractAttention.__init__`, and put it in `AbstractAttention.apply_causal_mask`. The causal mask is constructed at inference time, and its shape is `(cur_ctx_length, cur_ctx_length)`, which is generally *much* smaller than `(n_ctx, n_ctx)`.

I think the causal mask is light-weight enough that this should not cause performance issues. However, there is opportunity for further optimization: we could have a causal mask buffer tied to each instantiation of `AbstractAttention`, initialize it with some shape (maybe `(128, 128)`), and then increase it as needed (i.e. if we're doing a forward pass and the ctx_len is 129, then we can regenerate the buffer to be shape `(256, 256)`). This solution would avoid constructing new masks each forward pass. I didn't implement this here, but we can explore it if we feel it's necessary.

Fixes #479 

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# Checklist:

- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] I have not rewritten tests relating to key interfaces which would affect backward compatibility