@alan-cooney 2024-01-28
Also pinged you directly with a potential hacky (buy more efficient) fix using a static property