@andyrdt 2024-04-09
Hi @bryce13950 - thanks for pinging on this.

The currently-implemented solution in this PR is to construct attention masks for each attention component (i.e. at each layer) on-the-fly. This solution is simple, safe, and doesn't impact perf much (see benchmarks above), but feels a bit hacky since we're reconstructing/reallocating the same mask across many layers.

I think probably the best solution would be to construct a single attention mask at the model level at the beginning of a forward pass, and then pass this attention mask around. This is what I see most recent HuggingFace model implementations do.

For example, the HF Llama implementation [constructs](https://github.com/huggingface/transformers/blob/6f7d5db58c7c149c75642b5a4647b5cbc6c55643/src/transformers/models/llama/modeling_llama.py#L1043) a mask at the model level inside of `forward()`, and then plumbs it through the `forward()` of subcomponents, eventually [passing it](https://github.com/huggingface/transformers/blob/6f7d5db58c7c149c75642b5a4647b5cbc6c55643/src/transformers/models/llama/modeling_llama.py#L354) to the attention component’s `forward(..., attention_mask, ...)`.

There is a complication here with certain models that don’t use the same attention mask for each layer (e.g. GPT-neo, which [has alternating local/global attention for each layer](https://huggingface.co/EleutherAI/gpt-neo-2.7B/blob/main/config.json)). One way to deal with this would be to pass the global attention mask around, and then modify it locally in attention components that are configured to use `local` attention.

Let me know what you think about this proposed solution. Also please feel free to take a stab at implementing it (either editing this PR or submitting a new one) - I don't think I'll have time to properly revisit this over the next couple of weeks.